+++ tests/Unit/LinkTest.php
@@ -42,12 +42,12 @@ class LinkTest extends TableListTestCase
         $table->addColumn('name')->sortByDefault()->useForDestroyConfirmation()->isLink();
         $table->render();
         $html = view('tablelist::tbody', ['table' => $table])->render();
-        $this->assertContains('<a href="' . $user->name . '" title="validation.attributes.name">', $html);
+        $this->assertContains('<a href="' . $user->name . '" title="' . $user->name . '">', $html);
     }
 
     public function testIsLinkStringHtml()
     {
-        $this->createUniqueUser();
+        $user = $this->createUniqueUser();
         $this->setRoutes(['users'], ['index']);
         $routes = [
             'index' => ['alias' => 'users.index', 'parameters' => []],
@@ -56,12 +56,12 @@ class LinkTest extends TableListTestCase
         $table->addColumn('name')->sortByDefault()->useForDestroyConfirmation()->isLink('test');
         $table->render();
         $html = view('tablelist::tbody', ['table' => $table])->render();
-        $this->assertContains('<a href="test" title="validation.attributes.name">', $html);
+        $this->assertContains('<a href="test" title="' . $user->name . '">', $html);
     }
 
     public function testIsLinkClosureHtml()
     {
-        $this->createUniqueUser();
+        $user = $this->createUniqueUser();
         $this->setRoutes(['users'], ['index']);
         $routes = [
             'index' => ['alias' => 'users.index', 'parameters' => []],
@@ -72,7 +72,21 @@ class LinkTest extends TableListTestCase
         });
         $table->render();
         $html = view('tablelist::tbody', ['table' => $table])->render();
-        $this->assertContains('<a href="url" title="validation.attributes.name">', $html);
+        $this->assertContains('<a href="url" title="' . $user->name . '">', $html);
+    }
+
+    public function testIsLinkWithDefaultValueHtml()
+    {
+        $user = $this->createUniqueUser();
+        $this->setRoutes(['users'], ['index']);
+        $routes = [
+            'index' => ['alias' => 'users.index', 'parameters' => []],
+        ];
+        $table = app(TableList::class)->setRoutes($routes)->setModel(User::class);
+        $table->addColumn('name')->sortByDefault()->useForDestroyConfirmation()->isLink();
+        $table->render();
+        $html = view('tablelist::tbody', ['table' => $table])->render();
+        $this->assertContains('<a href="' . $user->name . '" title="' . $user->name . '">', $html);
     }
 
     public function testIsLinkWithNoValueHtml()
@@ -87,7 +101,7 @@ class LinkTest extends TableListTestCase
         $table->addColumn('name')->sortByDefault()->useForDestroyConfirmation()->isLink();
         $table->render();
         $html = view('tablelist::tbody', ['table' => $table])->render();
-        $this->assertNotContains('<a href="' . $user->name . '" title="validation.attributes.name">', $html);
+        $this->assertNotContains('<a href="', $html);
     }
 
     public function testIsLinkWithNoValueWithIconHtml()
@@ -102,6 +116,40 @@ class LinkTest extends TableListTestCase
         $table->addColumn('name')->sortByDefault()->useForDestroyConfirmation()->isLink()->setIcon('icon', true);
         $table->render();
         $html = view('tablelist::tbody', ['table' => $table])->render();
-        $this->assertContains('<a href="' . $user->name . '" title="validation.attributes.name">', $html);
+        $this->assertNotContains('<a href="', $html);
+    }
+
+    public function testIsLinkWithCustomValueHtml()
+    {
+        $user = $this->createUniqueUser();
+        $user->update(['name' => null]);
+        $this->setRoutes(['users'], ['index']);
+        $routes = [
+            'index' => ['alias' => 'users.index', 'parameters' => []],
+        ];
+        $table = app(TableList::class)->setRoutes($routes)->setModel(User::class);
+        $table->addColumn('name')->sortByDefault()->useForDestroyConfirmation()->isLink('url')->isCustomValue(function(){
+            return 'test';
+        });
+        $table->render();
+        $html = view('tablelist::tbody', ['table' => $table])->render();
+        $this->assertContains('<a href="url" title="test">', $html);
+    }
+
+    public function testIsLinkWithNoValueCustomValueHtml()
+    {
+        $user = $this->createUniqueUser();
+        $user->update(['name' => null]);
+        $this->setRoutes(['users'], ['index']);
+        $routes = [
+            'index' => ['alias' => 'users.index', 'parameters' => []],
+        ];
+        $table = app(TableList::class)->setRoutes($routes)->setModel(User::class);
+        $table->addColumn('name')->isLink()->isCustomValue(function(){
+            return 'test';
+        });
+        $table->render();
+        $html = view('tablelist::tbody', ['table' => $table])->render();
+        $this->assertContains('<a href="test" title="test">', $html);
     }
 }